// src/shared/api/models/getApiCampaigns200CampaignsItem.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * API
 * OpenAPI spec version: 1.0.0
 */
import type { GetApiCampaigns200CampaignsItemUpdatedAt } from './getApiCampaigns200CampaignsItemUpdatedAt';
import type { GetApiCampaigns200CampaignsItemCampaignStatus } from './getApiCampaigns200CampaignsItemCampaignStatus';

export type GetApiCampaigns200CampaignsItem = {
    id: string;
    name: string;
    customerGroup: string;
    isActive?: boolean;
    createdAt: string;
    updatedAt?: GetApiCampaigns200CampaignsItemUpdatedAt;
    openRate: string;
    recipients: number;
    revenue: string;
    campaignStatus: GetApiCampaigns200CampaignsItemCampaignStatus;
    /** Per-SMS charge when campaign type is SMS (0.70), otherwise 0 */
    perSmsCharge: number;
    /** Per-email charge when campaign type is SMS (0.00), otherwise 0 */
    perEmailCharge: number;
};